Added version text
I can also block any version from being played so thats good. Also if
you edit the source to bypass it I'll just have it so the client version
gets sent to the server for every request and if the version isnt valid
then it just sends back an error. Same for if you remove the version
parameter on the client.